requirements.txt
flask==3.0.3
python-dotenv==1.0.1

openai==1.35.3

langchain==0.2.11
langchain-core==0.2.26
langchain-community==0.2.10
langchain-openai==0.1.20

chromadb==0.5.5

pypdf==4.3.1
tiktoken==0.7.0

unstructured==0.15.7

boto3==1.34.162

sentence-transformers==3.0.1